NamNS: - Refactor package lists to sheets
