+++ src/main/java/com/prc391/patra/sheets/ListService.java
@@ -1,4 +1,4 @@
-package com.prc391.patra.lists;
+package com.prc391.patra.sheets;
 
 import com.prc391.patra.exceptions.EntityNotFoundException;
 import com.prc391.patra.tasks.Task;
